fix issue with wrongly cached image tags (#998)
